Add Github workflow for building documentation
This commit introduces a new Github workflow that is responsible for building documentation. This workflow automatically triggers on pushes to the "development" branch and a manual dispatch option is also available. The workflow makes use of the Writerside Docker builder to compile the documentation and, after building, saves the resulting artifacts to a location accessible for 7 days.

Signed-off-by: Akama Aka <rology@satowa-network.eu>